set authentication , others can't see thier notes

+++ app/Http/Controllers/NoteController.php
@@ -12,7 +12,10 @@ class NoteController extends Controller
      */
     public function index()
     {
-        $notes = Note::query()->orderBy("created_at","desc")->paginate();
+        $notes = Note::query()
+            ->where('user_id', request()->user()->id)
+            ->orderBy("created_at", "desc")
+            ->paginate();
 
         return view('note.index', ['notes' => $notes]);
     }
@@ -31,13 +34,13 @@ class NoteController extends Controller
     public function store(Request $request)
     {
         $data = $request->validate([
-            'note' => ['required','string']
+            'note' => ['required', 'string']
         ]);
 
-        $data['user_id'] = 1;
+        $data['user_id'] = $request->user()->id;
         $note = Note::create($data);
 
-        return to_route('note.show',$note)->with('message', 'Note was created');
+        return to_route('note.show', $note)->with('message', 'Note was created');
     }
 
     /**
@@ -45,7 +48,11 @@ class NoteController extends Controller
      */
     public function show(Note $note)
     {
-        return view('note.show', ['note'=> $note]);
+        if(request()->user()->id != $note->user_id) {
+            abort(403);   
+        }
+        
+        return view('note.show', ['note' => $note]);
     }
 
     /**
@@ -53,6 +60,10 @@ class NoteController extends Controller
      */
     public function edit(Note $note)
     {
+        if(request()->user()->id != $note->user_id) {
+            abort(403);   
+        }
+
         return view('note.edit', ['note' => $note]);
     }
 
@@ -61,12 +72,16 @@ class NoteController extends Controller
      */
     public function update(Request $request, Note $note)
     {
+        if(request()->user()->id != $note->user_id) {
+            abort(403);   
+        }
+
         $data = $request->validate([
-            'note'=> ['required','string']
+            'note' => ['required', 'string']
         ]);
 
         $note->update($data);
-        return to_route('note.show',$note)->with('message','Note was updated');
+        return to_route('note.show', $note)->with('message', 'Note was updated');
     }
 
     /**
@@ -74,7 +89,11 @@ class NoteController extends Controller
      */
     public function destroy(Note $note)
     {
+        if(request()->user()->id != $note->user_id) {
+            abort(403);   
+        }
+
         $note->delete();
-        return to_route('note.index')->with('message','Note was deleted');
+        return to_route('note.index')->with('message', 'Note was deleted');
     }
 }
